Added Error Types
